fix: separate tests for env initialization
