+++ src/middleware/logger/no-color.test.ts
@@ -1,0 +1,89 @@
+import { Hono } from '../../hono'
+
+describe('Logger by Middleware in NO_COLOR', () => {
+  let app: Hono
+  let log: string
+  let setLoggerColorEnabled: (enabled: boolean) => void
+
+  beforeEach(async () => {
+    vi.stubEnv('NO_COLOR', '1')
+    function sleep(time: number) {
+      return new Promise((resolve) => setTimeout(resolve, time))
+    }
+    const { logger, setLoggerColorEnabled: _ } = await import('.')
+    setLoggerColorEnabled = _
+
+    app = new Hono()
+
+    const logFn = (str: string) => {
+      log = str
+    }
+
+    const shortRandomString = 'hono'
+    const longRandomString = 'hono'.repeat(1000)
+
+    app.use('*', logger(logFn))
+    app.get('/short', (c) => c.text(shortRandomString))
+    app.get('/long', (c) => c.text(longRandomString))
+    app.get('/seconds', async (c) => {
+      await sleep(1000)
+
+      return c.text(longRandomString)
+    })
+    app.get('/empty', (c) => c.text(''))
+  })
+  afterAll(() => {
+    vi.unstubAllEnvs()
+  })
+  it('Log status 200 with empty body', async () => {
+    const res = await app.request('http://localhost/empty')
+    expect(res).not.toBeNull()
+    expect(res.status).toBe(200)
+    expect(log.startsWith('--> GET /empty 200')).toBe(true)
+    expect(log).toMatch(/m?s$/)
+  })
+
+  it('Log status 200 with small body', async () => {
+    const res = await app.request('http://localhost/short')
+    expect(res).not.toBeNull()
+    expect(res.status).toBe(200)
+    expect(log.startsWith('--> GET /short 200')).toBe(true)
+    expect(log).toMatch(/m?s$/)
+  })
+
+  it('Log status 200 with big body', async () => {
+    const res = await app.request('http://localhost/long')
+    expect(res).not.toBeNull()
+    expect(res.status).toBe(200)
+    expect(log.startsWith('--> GET /long 200')).toBe(true)
+    expect(log).toMatch(/m?s$/)
+  })
+
+  it('Time in seconds', async () => {
+    const res = await app.request('http://localhost/seconds')
+    expect(res).not.toBeNull()
+    expect(res.status).toBe(200)
+    expect(log.startsWith('--> GET /seconds 200')).toBe(true)
+    expect(log).toMatch(/1s/)
+  })
+
+  it('Log status 404', async () => {
+    const msg = 'Default 404 Not Found'
+    app.all('*', (c) => {
+      return c.text(msg, 404)
+    })
+    const res = await app.request('http://localhost/notfound')
+    expect(res).not.toBeNull()
+    expect(res.status).toBe(404)
+    expect(log.startsWith('--> GET /notfound 404')).toBe(true)
+    expect(log).toMatch(/m?s$/)
+  })
+
+  it('setLoggerColorEnabled take precedence over NO_COLOR', async () => {
+    setLoggerColorEnabled(true)
+    const res = await app.request('http://localhost/empty')
+    expect(res).not.toBeNull()
+    expect(res.status).toBe(200)
+    expect(log.startsWith('--> GET /empty \x1b[32m200\x1b[0m')).toBe(true)
+  })
+})
